+++ packages/platform/libraries/repositories.ts
@@ -1,7 +1,7 @@
 export { PrismaOOORepository } from "@calcom/lib/server/repository/ooo";
 export { ScheduleRepository as PrismaScheduleRepository } from "@calcom/lib/server/repository/schedule";
 export { BookingRepository as PrismaBookingRepository } from "@calcom/lib/server/repository/booking";
-export { EventTypeRepository as PrismaEventTypeRepository } from "@calcom/lib/server/repository/eventType";
+export { EventTypeRepository as PrismaEventTypeRepository } from "@calcom/lib/server/repository/eventTypeRepository";
 export { RoutingFormResponseRepository as PrismaRoutingFormResponseRepository } from "@calcom/lib/server/repository/formResponse";
 export { SelectedSlotsRepository as PrismaSelectedSlotsRepository } from "@calcom/lib/server/repository/selectedSlots";
 export { TeamRepository as PrismaTeamRepository } from "@calcom/lib/server/repository/team";
